book_nook/requirements.txt
fastapi
uvicorn[standard]
motor
passlib
python-dotenv
email-validator
beanie==1.26.0
resend
python-dotenv
pyjwt
pytest
httpx
bcrypt
